Rename islands folder hydration
